confirm event interface moved to event package
